benchmarks/pyproject.toml
[project]
name = "benchmarks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.5",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "psutil>=7.0.0",
    "tqdm>=4.67.1",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"
